old-tests/submission/W3C/domhtml/HTMLTableCaptionElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - TABLECAPTION</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-caption-align" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<table summary="Table Summary">
<caption align="top">CAPTION 1</caption>
<tbody><tr>
<th>Employee Id</th>
<th>Employee Name</th>
<th>Position</th>
<th>Salary</th>
</tr>
</tbody></table>









<script>
var domtest = async_test("The align IDL attribute of the caption element must reflect the content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var valign;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("caption");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      valign = testNode.align;

      assert_equals(valign,"top");
       
    domtest.done();
}));

</script>
<p>The align IDL attribute of the caption element must reflect the content attribute of the same name.</p>
<div id="log"></div>
</body></html>
